move interfaces to dedicated files
