+++ src/interfaces/IAWSConfig.ts
@@ -1,0 +1,6 @@
+export interface IAWSConfig {
+	readonly BUCKET_NAME: string;
+	readonly IDENTITY_POOL_ID: string;
+	readonly REGION: string;
+	readonly API_VERSION: string;
+}
