+++ src/interfaces/IImageUrlMapping.ts
@@ -1,0 +1,4 @@
+export interface IImageUrlMapping {
+	id: string;
+	url: string;
+}
