+++ tsconfig.json
@@ -1,5 +1,12 @@
 {
 	"compilerOptions": {
+		"baseUrl": "src",
+		"paths": {
+			"@components/*": ["components/*"],
+			"@interfaces/*": ["interfaces/*"],
+			"@services/*": ["services/*"],
+			"@views/*": ["views/*"]
+		},
 		"target": "es5",
 		"lib": ["dom", "dom.iterable", "esnext"],
 		"allowJs": true,
